Add files via upload
Features: The ID number of the person coming to the queue is requested. If
If ID=1, the ordinary customer is added to the end of the queue normally.
ID=2 has semi-priority because the bank has private customers. It goes 2 behind the person at the head of the queue.
If ID=3, it goes to the top of the queue because it is over 65 years old.